refactor: move evidence, scoring, and standards into core/ domain packages
Phase 4 of the clean architecture migration.

- engine/evidence.py → core/evidence/model.py
- engine/evidence_parser.py → core/evidence/parser.py
- engine/_merge.py → core/evidence/merge.py
- engine/scoring.py → core/scoring/engine.py
- engine/scoring_internals.py → core/scoring/internals.py
- engine/_scoring_numerical.py → core/scoring/numerical.py
- engine/_runner_report.py → core/scoring/report.py
- engine/standards.py → core/standards/loader.py
- engine/_ref_utils.py → core/standards/refs.py

Re-exports in old engine/ locations for backward compatibility.
