+++ src/quodeq/core/evidence/__init__.py
@@ -1,0 +1,2 @@
+from quodeq.core.evidence.model import Evidence, PrincipleEvidence, Judgment
+from quodeq.core.evidence.merge import merge_evidence
